+++ src/Led.cpp
@@ -348,10 +348,10 @@ static void Led_Task(void *parameter) {
 			nextAnimation = LedAnimationType::Idle;
 		} else if (gPlayProperties.pausePlay && !gPlayProperties.isWebstream) {
 			nextAnimation = LedAnimationType::Pause;
-		} else if (gPlayProperties.isWebstream) { // also animate pause in the webstream animation
-			nextAnimation = LedAnimationType::Webstream;
-		} else if ((gPlayProperties.playMode != BUSY) && (gPlayProperties.playMode != NO_PLAYLIST)) {
+		} else if ((gPlayProperties.playMode != BUSY) && (gPlayProperties.playMode != NO_PLAYLIST) && gPlayProperties.audioFileSize > 0) { // progress for a file/stream with known size
 			nextAnimation = LedAnimationType::Progress;
+		} else if (gPlayProperties.isWebstream) { // webstream animation (for streams with unknown size); pause animation is also handled by the webstream animation function
+			nextAnimation = LedAnimationType::Webstream;
 		} else if (gPlayProperties.playMode == NO_PLAYLIST) {
 			nextAnimation = LedAnimationType::Idle;
 		} else if (gPlayProperties.playMode == BUSY) {
@@ -1132,9 +1132,9 @@ AnimationReturnType Animation_BatteryMeasurement(const bool startNewAnimation, C
 
 		// fill all needed LEDs
 		if (filledLedCount < numLedsToLight) {
-			if (staticBatteryLevel < 0.3) {
+			if (staticBatteryLevel < 0.3f) {
 				leds[Led_Address(filledLedCount)] = CRGB::Red;
-			} else if (staticBatteryLevel < 0.6) {
+			} else if (staticBatteryLevel < 0.6f) {
 				leds[Led_Address(filledLedCount)] = CRGB::Orange;
 			} else {
 				leds[Led_Address(filledLedCount)] = CRGB::Green;